1.1.7

+++ src/main/java/vavix/io/FastByteArrayInputStream.java
@@ -67,7 +67,7 @@ public class FastByteArrayInputStream extends InputStream {
             n = count - pos;
         if (n < 0)
             return 0;
-        pos += n;
+        pos += (int) n;
         return n;
     }
 }
